Move RestrictedCardType to helper package.
COSDK-1204
